Move dns-control test to integration test
